+++ src/Workspaces/Core/Portable/LanguageServices/SemanticsFactsService/ISemanticFactsService.cs
@@ -5,8 +5,6 @@ using System.Collections.Immutable;
 using System.Linq;
 using System.Threading;
 using Microsoft.CodeAnalysis.Host;
-using Microsoft.CodeAnalysis.Shared.Extensions;
-using Microsoft.CodeAnalysis.Shared.Utilities;
 
 namespace Microsoft.CodeAnalysis.LanguageServices
 {
@@ -108,28 +106,5 @@ namespace Microsoft.CodeAnalysis.LanguageServices
         SyntaxToken GenerateUniqueName(
             SemanticModel semanticModel, SyntaxNode location, 
             SyntaxNode containerOpt, string baseName, CancellationToken cancellationToken);
-    }
-
-    internal abstract class AbstractSemanticFactsService
-    {
-        protected abstract ISyntaxFactsService SyntaxFactsService { get; }
-
-        public SyntaxToken GenerateUniqueName(
-            SemanticModel semanticModel, SyntaxNode location, SyntaxNode containerOpt,
-            string baseName, CancellationToken cancellationToken)
-        {
-            var syntaxFacts = this.SyntaxFactsService;
-
-            var container = containerOpt ?? location.AncestorsAndSelf().FirstOrDefault(
-                a => syntaxFacts.IsExecutableBlock(a) || syntaxFacts.IsMethodBody(a));
-            var existingSymbols = semanticModel.GetExistingSymbols(container, cancellationToken);
-
-            var reservedNames = semanticModel.LookupSymbols(location.SpanStart)
-                                             .Select(s => s.Name)
-                                             .Concat(existingSymbols.Select(s => s.Name));
-
-            return syntaxFacts.ToIdentifierToken(
-                NameGenerator.EnsureUniqueness(baseName, reservedNames, syntaxFacts.IsCaseSensitive));
-        }
     }
 }
